requirements.txt
flask
werkzeug
numpy
scikit-learn
pillow
pytesseract
chart.js